refactor: extract RetryPostTranscriptionResultHandler.swift from PostTranscriptionResultHandlers.swift
Splits the `RetryPostTranscriptionResultHandler` @MainActor class
(~61 lines) out of PostTranscriptionResultHandlers.swift into its own
file so that file holds only the `FinishedRecordingPostTranscriptionResultHandler`.

Byte-identical move of lines 79-139. Both files import Foundation.

PostTranscriptionResultHandlers.swift: 139 → 77 lines. Tests: 667.
